+++ Sources/KeyboardKit/Views/Emojis/EmojiKeyboard.swift
@@ -13,9 +13,11 @@ import SwiftUI
  This view can be used to list an emoji collection using the
  provided configuration.
  
- You can customize the buttons in the grid by using a custom
- `emojiButton` view builder function. You can also customize
- the button taps when using the standard builder function.
+ You can customize the emoji views in the keyboard, by using
+ the `emojiButton` initializer, which lets you customize the
+ keyboard button for every emoji. If you use the initializer
+ without a view builder, you will get a standard button with
+ the standard keyboard gestures for every emoji.
  */
 @available(iOS 14.0, tvOS 14.0, *)
 public struct EmojiKeyboard<ButtonView: View>: View {
@@ -26,23 +28,27 @@ public struct EmojiKeyboard<ButtonView: View>: View {
      - Parameters:
        - emojis: The emojis to include in the menu.
        - style: The style to apply to the keyboard, by default ``EmojiKeyboardStyle/standardPhonePortrait``.
+       - applyGestures: Whether or not to apply standard keyboard gestures to each button, by default `false`.
        - emojiButton: A emoji keyboard button builder function.
      */
     public init(
         emojis: [Emoji],
         style: EmojiKeyboardStyle = .standardPhonePortrait,
+        applyGestures: Bool = false,
         emojiButton: @escaping EmojiButtonBuilder<ButtonView>
     ) {
         let gridItem = GridItem(.fixed(style.itemSize), spacing: style.verticalItemSpacing - 9)
         self.emojis = emojis
         self.style = style
         self.rows = Array(repeating: gridItem, count: style.rows)
+        self.applyGestures = applyGestures
         self.emojiButton = emojiButton
     }
     
     private let emojis: [Emoji]
     private let rows: [GridItem]
     private let style: EmojiKeyboardStyle
+    private let applyGestures: Bool
     private let emojiButton: EmojiButtonBuilder<ButtonView>
     
     /**
@@ -53,14 +59,27 @@ public struct EmojiKeyboard<ButtonView: View>: View {
 
     public var body: some View {
         LazyHGrid(rows: rows, spacing: style.horizontalItemSpacing) {
-            ForEach(emojis) {
-                emojiButton($0, style)
-                    .accessibilityLabel($0.unicodeName ?? "")
-                    .accessibilityIdentifier($0.unicodeIdentifier ?? "")
+            ForEach(emojis) { emoji in
+                if applyGestures {
+                    buttonView(for: emoji, style: style)
+                        .withKeyboardGestures(for: .emoji(emoji), actionHandler: Self.standardKeyboardActionHandler)
+                } else {
+                    buttonView(for: emoji, style: style)
+                }
             }
         }
         .padding(.horizontal)
         .frame(height: style.totalHeight)
+    }
+}
+
+@available(iOS 14.0, tvOS 14.0, *)
+private extension EmojiKeyboard {
+
+    func buttonView(for emoji: Emoji, style: EmojiKeyboardStyle) -> some View {
+        emojiButton(emoji, style)
+            .accessibilityLabel(emoji.unicodeName ?? "")
+            .accessibilityIdentifier(emoji.unicodeIdentifier ?? "")
     }
 }
 
@@ -71,20 +90,35 @@ public extension EmojiKeyboard {
      This typealias represents an emoji-based action.
      */
     typealias EmojiAction = (Emoji) -> Void
-    
+
+    /**
+     The standard action handler to use to handle the emojis.
+     */
+    static var standardKeyboardActionHandler: KeyboardActionHandler {
+        KeyboardInputViewController.shared.keyboardActionHandler
+    }
+
     /**
      The standard action to use when tapping an emoji button.
      */
     static func standardEmojiAction(emoji: Emoji) {
-        let controller = KeyboardInputViewController.shared
-        let handler = controller.keyboardActionHandler
-        handler.handle(.tap, on: .emoji(emoji))
+        standardKeyboardActionHandler.handle(.tap, on: .emoji(emoji))
+    }
+
+    /**
+     The standard action to use when tapping an emoji button.
+     */
+    static func standardEmojiView(
+        for emoji: Emoji,
+        style: EmojiKeyboardStyle
+    ) -> some View {
+        EmojiKeyboardItem(emoji: emoji, style: style)
     }
 }
 
 @available(iOS 14.0, tvOS 14.0, *)
-public extension EmojiKeyboard where ButtonView == EmojiKeyboardButton {
-    
+public extension EmojiKeyboard {
+
     /**
      Create an emoji keyboard that uses standard buttons for
      each emoji in the provided collection.
@@ -97,8 +131,8 @@ public extension EmojiKeyboard where ButtonView == EmojiKeyboardButton {
     init(
         emojis: [Emoji],
         style: EmojiKeyboardStyle = .standardPhonePortrait,
-        emojiButtonAction: @escaping EmojiAction = standardEmojiAction
-    ) {
+        emojiButtonAction: @escaping EmojiAction
+    ) where ButtonView == EmojiKeyboardButton {
         self.init(
             emojis: emojis,
             style: style,
@@ -108,6 +142,26 @@ public extension EmojiKeyboard where ButtonView == EmojiKeyboardButton {
                     style: $1,
                     action: emojiButtonAction)
             }
+        )
+    }
+
+    /**
+     Create an emoji keyboard that uses standard buttons for
+     each emoji in the provided collection.
+
+     - Parameters:
+       - emojis: The emojis to include in the menu.
+       - style: The style to apply to the keyboard, by default ``EmojiKeyboardStyle/standardPhonePortrait``.
+     */
+    init(
+        emojis: [Emoji],
+        style: EmojiKeyboardStyle = .standardPhonePortrait
+    ) where ButtonView == EmojiKeyboardItem {
+        self.init(
+            emojis: emojis,
+            style: style,
+            applyGestures: true,
+            emojiButton: { EmojiKeyboardItem(emoji: $0, style: $1) }
         )
     }
 }